use local instead of global namespace
A Lua convention is to use the local namespace for (new; there is a deprecated/discouraged approach used in older versions of Lua) libraries and return a dictionary. A script using this library would then do `local lualib = require("lualib")` (or similar) and could then access the library's functions with `lualib.toHex`.